Fix mmap past file end on arm64 ios.
On arm64, where page size is 16k, if trying to map after the file end mmap doesn't map us the full page but only 4kb. This bug is present on iOS 8.2. As a workaround, we forcefully extend the file to the desired size to make sure than we can map the full page.

https://bugzilla.xamarin.com/show_bug.cgi?id=27667